DDF-3297 Tightens up default security policy and fixes cross-platform slash issue. (#2373)
This also cleans up the setenv scripts for Mac/Linux and Windows and includes a set of properties that can be used to determine missing security permissions. Runs the ProGrade Security Policy Security Manager.

Adds a permissions file to the `platform-osgi-configadmin` bundle to expand its rights.

Adds optional requirements to the pom for the `platform-osgi-configadmin` bundle.

Switches back to Felix configadmin in favor of DDF-override.

Runs the ConfigurationInstaller's code in a privileged block.

Updates itests to run with default.policy.

Adds option to itests for generated missing policy file.